moves ActiveTodosCount to todos feature
