--- components/active-todos-count/index.ts
@@ -1,1 +1,0 @@
-export * from './active-todos-count'
