+++ features/todos/components/active-todos-count.tsx
@@ -1,0 +1,14 @@
+import styles from '../todos.module.scss'
+import { useActiveTodosCount } from '../queries'
+
+export function ActiveTodosCount() {
+  const activeTodosCountQuery = useActiveTodosCount()
+
+  if (!activeTodosCountQuery.isSuccess) return null
+
+  const itemsLeftText = `${activeTodosCountQuery.data} item${
+    activeTodosCountQuery.data !== 1 ? 's' : ''
+  } left`
+
+  return <div className={styles.itemsCount}>{itemsLeftText}</div>
+}
