+++ features/todos/todos.module.scss
@@ -172,3 +172,12 @@
     padding-right: var(--s1);
   }
 }
+
+.itemsCount {
+  composes: smallText;
+
+  grid-area: count;
+  border-bottom-left-radius: 4px;
+  padding: var(--s0) var(--s1);
+  color: var(--muted-text-color);
+}
